requirements.txt
fastapi
uvicorn
sqlalchemy
requests
beautifulsoup4
python-dotenv
apscheduler
fake-useragent
streamlit
pandas
plotly
